// src/src/app/tramites/mintransporte/licencia-conduccion/models/SolicitudCertificadoModel.ts
export interface SolicitudCertificadoModel {
        PrimerNombre: string;

        SegundoNombre: string;

        PrimerApellido: string;

        SegundoApellido: string; 

        TipoDocumento: string;

        NumIdentificacion: string;

        CorreoElectronico: string ;

        TipoEntidad: string;

        Pais: string ;
}